wetube

+++ controllers/videoController.js
@@ -1,6 +1,8 @@
 import routes from "../routes";
 import Video from "../models/Video";
 
+// home
+
 export const home = async (req, res) => {
   try {
     const videos = await Video.find({}).sort({ _id: -1 });
@@ -10,6 +12,8 @@ export const home = async (req, res) => {
     res.render("home", { pageTitle: "home", videos: [] });
   }
 };
+
+// search
 
 export const search = async (req, res) => {
   const {
@@ -26,8 +30,10 @@ export const search = async (req, res) => {
   res.render("search", { pageTitle: "search", searchingBy, videos });
 };
 
+// Upload Video
 export const getUpload = (req, res) =>
   res.render("upload", { pageTitle: "upload" });
+
 export const postUpload = async (req, res) => {
   const {
     body: { title, description },
@@ -36,18 +42,22 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
-  console.log(newVideo);
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
+
+// Video Detail
 
 export const videoDetail = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -55,17 +65,24 @@ export const videoDetail = async (req, res) => {
   }
 };
 
+// Edit Video
+
 export const getEditVideo = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
 };
+
 export const postEditVideo = async (req, res) => {
   const {
     params: { id },
@@ -79,12 +96,19 @@ export const postEditVideo = async (req, res) => {
   }
 };
 
+// Delete Video
+
 export const deleteVideo = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
